+++ acme/acme/messages.py
@@ -12,6 +12,7 @@ from typing import Optional
 from typing import Tuple
 from typing import Type
 from typing import TypeVar
+import warnings
 
 import josepy as jose
 
@@ -233,6 +234,7 @@ class Directory(jose.JSONDeSerializable):
         website: str = jose.field('website', omitempty=True)
         caa_identities: List[str] = jose.field('caaIdentities', omitempty=True)
         external_account_required: bool = jose.field('externalAccountRequired', omitempty=True)
+        profiles: Dict[str, str] = jose.field('profiles', omitempty=True)
 
         def __init__(self, **kwargs: Any) -> None:
             kwargs = {self._internal_name(k): v for k, v in kwargs.items()}
@@ -584,7 +586,11 @@ class CertificateRequest(jose.JSONObjectWithFields):
         `OpenSSL.crypto.X509Req` wrapped in `.ComparableX509`
 
     """
-    csr: jose.ComparableX509 = jose.field('csr', decoder=jose.decode_csr, encoder=jose.encode_csr)
+    with warnings.catch_warnings():
+        warnings.filterwarnings('ignore',
+            message='The next major version of josepy will remove josepy.util.ComparableX509')
+        csr: jose.ComparableX509 = jose.field(
+            'csr', decoder=jose.decode_csr, encoder=jose.encode_csr)
 
 
 class CertificateResource(ResourceWithURI):
@@ -596,7 +602,10 @@ class CertificateResource(ResourceWithURI):
     :ivar tuple authzrs: `tuple` of `AuthorizationResource`.
 
     """
-    cert_chain_uri: str = jose.field('cert_chain_uri')
+    with warnings.catch_warnings():
+        warnings.filterwarnings('ignore',
+            message='The next major version of josepy will remove josepy.util.ComparableX509')
+        cert_chain_uri: str = jose.field('cert_chain_uri')
     authzrs: Tuple[AuthorizationResource, ...] = jose.field('authzrs')
 
 
@@ -607,14 +616,19 @@ class Revocation(jose.JSONObjectWithFields):
         `jose.ComparableX509`
 
     """
-    certificate: jose.ComparableX509 = jose.field(
-        'certificate', decoder=jose.decode_cert, encoder=jose.encode_cert)
+    with warnings.catch_warnings():
+        warnings.filterwarnings('ignore',
+            message='The next major version of josepy will remove josepy.util.ComparableX509')
+        certificate: jose.ComparableX509 = jose.field(
+            'certificate', decoder=jose.decode_cert, encoder=jose.encode_cert)
     reason: int = jose.field('reason')
 
 
 class Order(ResourceBody):
     """Order Resource Body.
 
+    :ivar profile: The profile to request.
+    :vartype profile: str
     :ivar identifiers: List of identifiers for the certificate.
     :vartype identifiers: `list` of `.Identifier`
     :ivar acme.messages.Status status:
@@ -626,6 +640,8 @@ class Order(ResourceBody):
     :ivar datetime.datetime expires: When the order expires.
     :ivar ~.Error error: Any error that occurred during finalization, if applicable.
     """
+    # https://datatracker.ietf.org/doc/draft-aaron-acme-profiles/
+    profile: str = jose.field('profile', omitempty=True)
     identifiers: List[Identifier] = jose.field('identifiers', omitempty=True)
     status: Status = jose.field('status', decoder=Status.from_json, omitempty=True)
     authorizations: List[str] = jose.field('authorizations', omitempty=True)